+++ plugins/postcss-cascade-layers/src/constants.ts
@@ -1,0 +1,2 @@
+export const INVALID_LAYER_NAME = 'invalid-layer';
+export const WITH_SELECTORS_LAYER_NAME = 'layer-with-selector-rules';
