+++ plugins/postcss-cascade-layers/src/index.ts
@@ -1,11 +1,15 @@
-import type { Container, AtRule, Node, PluginCreator } from 'postcss';
 import selectorParser from 'postcss-selector-parser';
+import type { Container, AtRule, PluginCreator } from 'postcss';
+import { Model } from './model';
 import { adjustSelectorSpecificity } from './adjust-selector-specificity';
+import { desugarAndParseLayerNames } from './desugar-and-parse-layer-names';
 import { desugarNestedLayers } from './desugar-nested-layers';
 import { getLayerAtRuleAncestor } from './get-layer-atrule-ancestor';
-import { Model } from './model';
-import { someInTree } from './some-in-tree';
 import { selectorSpecificity } from './specificity';
+import { someAtRuleInTree } from './some-in-tree';
+import { sortRootNodes } from './sort-root-nodes';
+import { recordLayerOrder } from './record-layer-order';
+import { INVALID_LAYER_NAME } from './constants';
 
 const creator: PluginCreator<undefined> = () => {
 	return {
@@ -13,150 +17,13 @@ const creator: PluginCreator<undefined> = () => {
 		Once(root: Container) {
 			const model = new Model();
 
-			// - parse layer names
-			// - rename anon layers
-			// - handle empty layers
-			root.walkAtRules('layer', (layerRule) => {
-				if (layerRule.params) {
-					const layerNameList: Array<string> = [];
-					let isInvalidLayerName = false;
+			desugarAndParseLayerNames(root, model);
 
-					selectorParser().astSync(layerRule.params).each((selector) => {
-						const currentLayerNameParts: Array<string> = [];
-
-						selector.walk((node) => {
-							switch (node.type) {
-								case 'class':
-									currentLayerNameParts.push(node.value);
-									break;
-								case 'tag':
-									currentLayerNameParts.push(node.value);
-									break;
-								default:
-									isInvalidLayerName = true;
-									break;
-							}
-						});
-
-						if (isInvalidLayerName) {
-							return;
-						}
-
-						layerNameList.push(currentLayerNameParts.join('.'));
-
-						model.addLayerNameParts(currentLayerNameParts);
-					});
-
-					model.addLayerParams(layerRule.params, layerNameList);
-
-					if (layerRule.nodes && layerNameList.length > 1) {
-						// If the layer is a container rule it can not have multiple layer names.
-						isInvalidLayerName = true;
-					}
-
-					if (isInvalidLayerName) {
-						// Set invalid layers to "invalid-layer"
-						// We reset these later.
-						layerRule.name = 'invalid-layer';
-						return;
-					}
-
-					// handle empty layer at-rules.
-					if (!layerRule.nodes || layerRule.nodes.length === 0) {
-						layerNameList.forEach((name) => {
-							model.getLayerNameList(name).forEach((part) => {
-								if (model.layerOrder.has(part)) {
-									return;
-								}
-
-								model.layerOrder.set(part, model.layerCount);
-								model.layerCount += 1;
-							});
-						});
-
-						layerRule.remove();
-						return;
-					}
-				}
-
-				// give anonymous layers a name
-				if (!layerRule.params) {
-					layerRule.raws.afterName = ' ';
-					layerRule.params = model.createAnonymousLayerName();
-				}
-
-				let hasNestedLayers = false;
-				let hasUnlayeredStyles = false;
-
-				// check for where a layer has nested layers AND styles outside of those layers
-				layerRule.each((node) => {
-					if (node.type === 'atrule' && node.name === 'layer') {
-						hasNestedLayers = true;
-					} else {
-						hasUnlayeredStyles = true;
-					}
-
-					if (hasNestedLayers && hasUnlayeredStyles) {
-						return false;
-					}
-				});
-
-				if (hasNestedLayers && hasUnlayeredStyles) {
-					// create new final layer via cloning and keep only the styles
-					const implicitLayerName = model.createImplicitLayerName(layerRule.params);
-					const implicitLayer = layerRule.clone({
-						params: implicitLayerName,
-					});
-
-					implicitLayer.each((node) => {
-						if (node.type === 'atrule' && node.name === 'layer') {
-							node.remove();
-						}
-					});
-
-					// insert new layer
-					layerRule.append(implicitLayer);
-
-					// go through the unlayered rules and delete from top level atRule
-					layerRule.each((node) => {
-						if (node.type === 'atrule' && node.name === 'layer') {
-							return;
-						}
-
-						node.remove();
-					});
-				}
-			});
-
-			// record layer order
-			root.walkAtRules('layer', (layerRule) => {
-				const currentLayerNameParts = model.getLayerParams(layerRule);
-				const fullLayerName = currentLayerNameParts.join('.');
-				if (model.layerOrder.has(fullLayerName)) {
-					return;
-				}
-
-				if (!model.layerParamsParsed.has(fullLayerName)) {
-					model.layerParamsParsed.set(fullLayerName, [fullLayerName]);
-				}
-
-				if (!model.layerNameParts.has(fullLayerName)) {
-					model.layerNameParts.set(fullLayerName, [...currentLayerNameParts]);
-				}
-
-				model.getLayerNameList(fullLayerName).forEach((name) => {
-					if (model.layerOrder.has(name)) {
-						return;
-					}
-
-					model.layerOrder.set(name, model.layerCount);
-					model.layerCount += 1;
-				});
-			});
+			recordLayerOrder(root, model);
 
 			if (!model.layerCount) {
 				// Reset "invalid-layer" at rules
-				root.walkAtRules('invalid-layer', (layerRule) => {
+				root.walkAtRules(INVALID_LAYER_NAME, (layerRule) => {
 					layerRule.name = 'layer';
 				});
 
@@ -176,12 +43,12 @@ const creator: PluginCreator<undefined> = () => {
 
 			// transform unlayered styles - need highest specificity (layerCount)
 			root.walkRules((rule) => {
-				if (getLayerAtRuleAncestor(rule)) {
+				// Skip any at rules that do not contain regular declarations (@keyframes)
+				if (rule.parent && rule.parent.type === 'atrule' && (rule.parent as AtRule).name === 'keyframes') {
 					return;
 				}
 
-				// Skip any at rules that do not contain regular declarations (@keyframes)
-				if (rule.parent && rule.parent.type === 'atrule' && (rule.parent as AtRule).name === 'keyframes') {
+				if (getLayerAtRuleAncestor(rule)) {
 					return;
 				}
 
@@ -199,79 +66,18 @@ const creator: PluginCreator<undefined> = () => {
 			// Transform order of CSS
 			// - split selector rules from  non-selector rules
 			// - sort non-selector rules
-			{
-				// Separate selector rules from other rules
-				root.walkAtRules('layer', (layerRule) => {
-					const withSelectorRules = layerRule.clone();
-					const withoutSelectorRules = layerRule.clone();
-
-					withSelectorRules.walkAtRules((atRule) => {
-						if (atRule.name === 'keyframes') {
-							const parent = atRule.parent;
-							atRule.remove();
-							if (parent.nodes.length === 0) {
-								parent.remove();
-							}
-
-							return;
-						}
-
-						if (someInTree(atRule, (node) => {
-							return node.type === 'rule';
-						})) {
-							return;
-						}
-
-						const parent = atRule.parent;
-						atRule.remove();
-						if (parent.nodes.length === 0) {
-							parent.remove();
-						}
-					});
-
-					withoutSelectorRules.walkRules((rule) => {
-						if (rule.parent && rule.parent.type === 'atrule' && (rule.parent as AtRule).name === 'keyframes') {
-							return;
-						}
-
-						const parent = rule.parent;
-						rule.remove();
-						if (parent.nodes.length === 0) {
-							parent.remove();
-						}
-					});
-
-					withSelectorRules.name = 'layer-with-selector-rules';
-
-					layerRule.replaceWith(withSelectorRules, withoutSelectorRules);
-					if (withSelectorRules.nodes.length === 0) {
-						withSelectorRules.remove();
-					}
-
-					if (withoutSelectorRules.nodes.length === 0) {
-						withoutSelectorRules.remove();
-					}
-				});
-
-				// Sort layer nodes
-				model.sortRootNodes(root.nodes);
-
-				// Reset "layer-with-selector-rules" at rules
-				root.walkAtRules('layer-with-selector-rules', (atRule) => {
-					atRule.name = 'layer';
-				});
-			}
+			sortRootNodes(root, model);
 
 			// transform layered styles:
 			//  - give selectors the specificity they need based on layerPriority state
 			root.walkRules((rule) => {
-				const layerForCurrentRule = getLayerAtRuleAncestor(rule);
-				if (!layerForCurrentRule) {
+				// Skip any at rules that do not contain regular declarations (@keyframes)
+				if (rule.parent && rule.parent.type === 'atrule' && (rule.parent as AtRule).name === 'keyframes') {
 					return;
 				}
 
-				// Skip any at rules that do not contain regular declarations (@keyframes)
-				if (rule.parent && rule.parent.type === 'atrule' && (rule.parent as AtRule).name === 'keyframes') {
+				const layerForCurrentRule = getLayerAtRuleAncestor(rule);
+				if (!layerForCurrentRule) {
 					return;
 				}
 
@@ -283,14 +89,14 @@ const creator: PluginCreator<undefined> = () => {
 
 			// Remove all @layer at-rules
 			// Contained styles are inserted before
-			while (someInTree(root, (node) => node.type === 'atrule' && node.name === 'layer')) {
+			while (someAtRuleInTree(root, (node) => node.name === 'layer')) {
 				root.walkAtRules('layer', (atRule) => {
 					atRule.replaceWith(atRule.nodes);
 				});
 			}
 
 			// Reset "invalid-layer" at rules
-			root.walkAtRules('invalid-layer', (atRule) => {
+			root.walkAtRules(INVALID_LAYER_NAME, (atRule) => {
 				atRule.name = 'layer';
 			});
 		},
